requirements.txt
Flask
Flask-Migrate
Flask-SQLAlchemy
Flask-Security-Too
flask_wtf
psycopg2
python-dotenv
SQLAlchemy
gmqtt
uvloop